info-bot/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
discord==1.0.1
discord.py==1.3.3
idna==2.9
multidict==4.7.5
newsapi-python==0.2.6
requests==2.23.0
soupsieve==2.0
urllib3==1.25.9
websockets==8.1
wikipedia==1.4.0
yarl==1.4.2
